[WIP] Add support for syncing components

+++ osfoffline/sync/remote.py
@@ -5,18 +5,31 @@ from pathlib import Path
 import threading
 import time
 
+from sqlalchemy.orm.exc import NoResultFound
+
+from pathlib import Path
+
 from osfoffline import settings
+
+from osfoffline.client.osf import OSFClient
+
 from osfoffline.database import Session
 from osfoffline.database.models import Node, File
+from osfoffline.database.utils import save
+
+from osfoffline.sync.local import LocalSyncWorker
 from osfoffline.sync.exceptions import FolderNotInFileSystem
-from osfoffline.sync.ext.auditor import Auditor
-from osfoffline.sync.ext.auditor import EventType
+from osfoffline.sync.ext.auditor import (
+    Auditor,
+    EventType
+)
+
 from osfoffline.tasks.notifications import Notification
 from osfoffline.tasks.resolution import RESOLUTION_MAP
+from osfoffline.tasks.queue import OperationWorker
+
 from osfoffline.utils import Singleton
 from osfoffline.utils.authentication import get_current_user
-from osfoffline.tasks.queue import OperationWorker
-from osfoffline.sync.local import LocalSyncWorker
 
 
 logger = logging.getLogger(__name__)
@@ -35,7 +48,7 @@ class RemoteSyncWorker(threading.Thread, metaclass=Singleton):
 
     def initialize(self):
         logger.info('Beginning initial sync')
-        for node in Session().query(Node).all():
+        for node in Session().query(Node).filter(Node.sync).all():
             self._preprocess_node(node)
         Session().commit()
         # TODO No need for this check
@@ -55,18 +68,18 @@ class RemoteSyncWorker(threading.Thread, metaclass=Singleton):
                 logger.info('Sleep interrupted, syncing now')
             self._sync_now_event.clear()
 
-            # Ensure selected node directories exist
-            for node in Session().query(Node).all():
-                local = Path(os.path.join(node.path, settings.OSF_STORAGE_FOLDER))
+            logger.info('Beginning remote sync')
+            LocalSyncWorker().ignore.set()
+
+            # Ensure selected node directories exist and db entries created
+            for node in Session().query(Node).filter(Node.sync).all():
                 try:
-                    os.makedirs(str(local), exist_ok=True)
+                    self._preprocess_node(node, delete=False)
                 except OSError:
                     # TODO: If the node folder cannot be created, what further actions must be taken before attempting to sync?
                     # TODO: Should the error be user-facing?
                     logger.exception('Error creating node directory for sync')
 
-            logger.info('Beginning remote sync')
-            LocalSyncWorker().ignore.set()
             OperationWorker().join_queue()
 
             try:
@@ -87,12 +100,58 @@ class RemoteSyncWorker(threading.Thread, metaclass=Singleton):
         self.__stop.set()
         self.sync_now()
 
-    def _preprocess_node(self, node):
-        local = Path(os.path.join(node.path, settings.OSF_STORAGE_FOLDER))
-        if not local.exists():
-            logger.warning('Clearing files for node {}'.format(node))
-            Session().query(File).filter(File.node_id == node.id).delete()
-        os.makedirs(str(local), exist_ok=True)
+    def _orphan_children(self, node, remote_children):
+        """It's a hard world out there...
+        Delete the database record for any descendant not mirrored remotely.
+        Via cascade this will also remove any descedant Nodes and Files.
+        The effect of this action is that any files associated with a child Node
+        locally for which the remote Node has been deleted are explicitly removed
+        from OSFO's auditing and will be ignored.
+        """
+        children_ids = map(lambda c: c.id, remote_children)
+        for record in node.children:
+            if record.id not in children_ids:
+                Session().delete(record)
+
+    def _preprocess_node(self, node, delete=True):
+        nodes = [node]
+        remote_node = OSFClient().get_node(node.id)
+        stack = remote_node.get_children(lazy=False)
+        self._orphan_children(node, stack)
+        while len(stack):
+            child = stack.pop(0)
+            # Ensure the database contains a Node record for each node in the project heirarchy.
+            # This must guarentee the remote/database representations of the project heirarchy are
+            # fully congruent.
+            # TODO: If we want to support syncing only subsets of the project heirarchy then some
+            # additional logic could be added here to skip over certain nodes.
+            try:
+                db_child = Session().query(Node).filter(
+                    Node.id == child.id
+                ).one()
+            except NoResultFound:
+                # Setting sync=False notes that the node is implicity synced
+                parent = Session().query(Node).filter(
+                    Node.id == child.parent.id
+                ).one()
+                db_child = Node(
+                    id=child.id,
+                    title=child.title,
+                    user=node.user,
+                    parent_id=parent.id
+                )
+                Session().add(db_child)
+            nodes.append(db_child)
+            children = child.get_children(lazy=False)
+            self._orphan_children(db_child, children)
+            stack = stack + children
+        save(Session())
+        for node in nodes:
+            local = Path(os.path.join(node.path, settings.OSF_STORAGE_FOLDER))
+            if delete and not local.exists():
+                logger.warning('Clearing files for node {}'.format(node))
+                Session().query(File).filter(File.node_id == node.id).delete()
+            os.makedirs(str(local), exist_ok=True)
 
     def _check(self):
         resolutions = []
